requirements.txt
aiogram>=3.10.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
